Fix two home page nits
We've a bad link to the conmon GitHub repo, that repairs that.  It was also
pointed out that we've not posted a blog to the medium/crio site in over two
years, this comments the link to that out.  If more blogs are added there (Yes Please!), then we can turn that back on.

Addresses: #46
Addresses: #48

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>